LayoutTests/media/unsupported-rtsp-expected.txt

Test that QuickTime file with RTSP URL generates a load error.

EVENT(error)
EXPECTED (video.error != 'null') OK
EXPECTED (video.error.code == '4') OK
EXPECTED (video.networkState == '3') OK
END OF TEST
